refactor: stack navigator 리팩토링
