+++ ReactNativeStudy/Breeze/Week2/CameraPictureApp/src/navigation/Navigation.tsx
@@ -1,11 +1,6 @@
 import {NavigationContainer, DefaultTheme} from '@react-navigation/native';
-import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import BottomTab from './BottomTab';
-import DetailScreen from '../screen/DetailScreen';
-import {RootStackParamList} from '../type/rootStack';
 import {palette} from '../styles/ColorPalette';
-
-const Stack = createNativeStackNavigator<RootStackParamList>();
+import StackNavigator from './StackNavigator';
 
 function Navigation() {
   const navTheme = {
@@ -18,16 +13,7 @@ function Navigation() {
 
   return (
     <NavigationContainer theme={navTheme}>
-      <Stack.Navigator initialRouteName={'BottomTab'}>
-        <Stack.Screen
-          name="BottomTab"
-          component={BottomTab}
-          options={{
-            headerShown: false,
-          }}
-        />
-        <Stack.Screen name="Detail" component={DetailScreen} />
-      </Stack.Navigator>
+      <StackNavigator />
     </NavigationContainer>
   );
 }
