+++ ReactNativeStudy/Breeze/Week2/CameraPictureApp/src/type/navigator.ts
@@ -1,7 +1,12 @@
 export type RootStackParamList = {
-  Detail: {img: string};
+  TabNavigator: undefined;
+  Detail: {
+    img: string;
+  };
+};
+
+export type TabParamList = {
   Home: undefined;
   Picture: undefined;
   Setting: undefined;
-  BottomTab: undefined;
 };
